fix: use bigint for risk_bucket to prevent integer overflow on INT_MIN
abs(-2147483648) overflows signed int32. Using bigint avoids this edge
case when the first 8 hex chars of the MD5 hash are exactly '80000000'.

+++ dbt_project/models/intermediate/int_customer_risk_factors.sql
@@ -83,7 +83,7 @@ risk_factors as (
         end as derived_risk_rating,
 
         -- Deterministic bucket (replaces HASHBUCKET(HASHROW(customer_id)) MOD 10)
-        mod(abs(('x' || substring({{ dbt_utils.generate_surrogate_key(['ca.customer_id']) }}, 1, 8))::bit(32)::int), 10) as risk_bucket
+        mod(abs(('x' || substring({{ dbt_utils.generate_surrogate_key(['ca.customer_id']) }}, 1, 8))::bit(32)::bigint), 10) as risk_bucket
 
     from customer_accounts ca
     left join transaction_metrics tm
